Created abstract class Notes, and Sticky Note class which extends it
